requirements.txt
streamlit
langchain
langchain-community
langchain-openai
langchain-core
python-dotenv
chromadb==0.4.15
protobuf>=4.21.6,<5.0.0
pypdf
boto3
requests
pysqlite3-binary